"Better menu and game balance"

+++ src/towerDefense/entity/enemies/EnemyClass.java
@@ -11,16 +11,16 @@ import towerDefense.ui.HealthBar;
 
 public abstract class EnemyClass extends EntityClass {
 
-    private int health;
-    private int maxHealth;
-    private int armor;
+    private double health;
+    private double maxHealth;
+    private double armor;
     private double speed;
     
     private int reward;
 
     private HealthBar healthBar = new HealthBar(this.getLayer(), this.getPosX(), this.getPosY() - 5);
 
-    protected EnemyClass(Pane layer, Image image, int health, int armor, double speed, int reward){
+    protected EnemyClass(Pane layer, Image image, double health, double armor, double speed, int reward){
         super(layer, image, Config.SPAWN_POS_X, Config.SPAWN_POS_Y, Config.SPAWN_ROTATION);
         this.health = health;
         this.maxHealth = health;
@@ -31,13 +31,13 @@ public abstract class EnemyClass extends EntityClass {
     
     // Getters & Setters
     //#region
-    public void setHealth(int health) {this.health = health;}
-    public void setArmor(int armor) {this.armor = armor;}
+    public void setHealth(double health) {this.health = health;}
+    public void setArmor(double armor) {this.armor = armor;}
     public void setSpeed(double speed) {this.speed = speed;}
     public void setReward (int reward) {this.reward = reward;}
     
-    public int getHealth() {return this.health;}
-    public int getArmor() {return this.armor;}
+    public double getHealth() {return this.health;}
+    public double getArmor() {return this.armor;}
     public double getSpeed() {return this.speed;}
     public int getReward() {return this.reward;}
